requirements.txt
behave==1.2.6
parse==1.19.0
parse-type==0.5.2
python-dotenv==0.19.0
selenium==3.141.0
six==1.16.0
urllib3==1.26.6
